Edit/add more fields

+++ src/main/resources/db/migrations/V1.3.1__create_loan_table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE loan
 (
     id bigint AUTO_INCREMENT PRIMARY KEY,
-    amount_to_repay float NOT NULL,
-    amount_payed float NOT NULL,
+    amount_to_repay DECIMAL(7,2) NOT NULL,
+    amount_payed DECIMAL(7,2) NOT NULL,
     status VARCHAR(255) NOT NULL,
     customer_id bigint NOT NULL,
     loan_application_id bigint NOT NULL,
